requirements.txt
Django==1.5.1
argparse==1.2.1
distribute==0.6.24
dj-database-url==0.2.1
gunicorn==0.17.4
psycopg2==2.5
pyclips.tar.gz
wsgiref==0.1.2
